awt & png renderer written. runs beautifully.
git-svn-id: https://svn.q1cc.net/repository/nixnick/buddha@474 a70f5666-fd2b-40d0-8f41-e30882f606c2

+++ src/buddha/Renderer.java
@@ -16,7 +16,7 @@ public interface Renderer {
      * @param y
      * @param value 
      */
-    void setPoint(int x,int y, float value);
+    void setPoint(int x,int y, double value);
     
     /**
      * adds a value to a specific point
@@ -24,7 +24,7 @@ public interface Renderer {
      * @param y
      * @param value
      */
-    void addToPoint(int x,int y, float value);
+    void addToPoint(int x,int y, double value);
     
     long getNumDataRecvd();
